added contactHistory / Created latestVersion
latestVersion is a copy of contact.history so that we don't change the original array when we want to edit our contact. latestVersion is the person that's always the "active" on. 
Created a table for the history and a forEach loop to go through the history array with their given index. on the resetButton I put in data-index so we know which place the contact have in the array.